requirements.txt

dash~=2.13.0
dash-bootstrap-components==0.13.0
dash-table==4.11.3
plotly==5.3.1
requests==2.26.0
mysql-connector-python==8.0.26
Flask~=2.2.5
PyMySQL~=1.1.0
Faker~=19.6.2
